Split library into multiple files
