--- django_multitenant/__init__.py
@@ -1,1 +1,0 @@
-from django_multitenant.django_multitenant import *
