--- django_multitenant/django_multitenant.py
@@ -1,248 +1,0 @@
-from django.apps import apps
-from django.db import models
-try:
-    from threading import local
-except ImportError:
-    from django.utils._threading_local import local
-
-from collections import OrderedDict
-import pdb
-
-_thread_locals = local()
-
-
-#Modified QuerySet to add suitable filters on joins.
-class TenantQuerySet(models.QuerySet):
-    
-    #This is adding tenant filters for all the models in the join.
-    def add_tenant_filters_with_joins(self):
-        current_tenant=get_current_tenant()
-        if current_tenant:
-            extra_sql=[]
-            extra_params=[]
-            current_table_name=self.model._meta.db_table
-            alias_refcount = self.query.alias_refcount
-            alias_map = self.query.alias_map
-            for k,v in alias_refcount.items():
-                if(v>0 and k!=current_table_name):
-                    current_model=get_model_by_db_table(alias_map[k].table_name)
-                    if issubclass(current_model, TenantModel):
-                        extra_sql.append(k+'."'+current_model.tenant_id+'" = %s')
-                        extra_params.append(current_tenant.id)
-            self.query.add_extra([],[],extra_sql,extra_params,[],[])
-    
-    def add_tenant_filters_without_joins(self):
-        current_tenant=get_current_tenant()
-        if current_tenant:
-            l=[]
-            table_name=self.model._meta.db_table
-            l.append(table_name+'.'+self.model.tenant_id+'='+str(current_tenant.id))
-            self.query.add_extra([],[],l,[],[],[])
-
-    #Below are APIs which generate SQL, this is where the tenant_id filters are injected for joins.
-    
-    def __iter__(self):
-        self.add_tenant_filters_with_joins()
-        return super(TenantQuerySet,self).__iter__()
-    
-    def aggregate(self, *args, **kwargs):
-        self.add_tenant_filters_with_joins()
-        return super(TenantQuerySet,self).aggregate(*args, **kwargs)
-
-    def count(self):
-        self.add_tenant_filters_with_joins()
-        return super(TenantQuerySet,self).count()
-
-    def get(self, *args, **kwargs):
-        #self.add_tenant_filters()
-        self.add_tenant_filters_with_joins()
-        return super(TenantQuerySet,self).get(*args,**kwargs)
-
-    # def get_or_create(self, defaults=None, **kwargs):
-    #     self.add_tenant_filters()
-    #     return super(TenantQuerySet,self).get_or_create(defaults,**kwargs)
-
-    # def update(self, **kwargs):
-    #     self.add_tenant_filters_without_joins()
-    #     #print(self.query.alias_refcount)
-    #     return super(TenantQuerySet,self).update(**kwargs)
-    
-    # def _update(self, values):
-    #     self.add_tenant_filters_without_joins()
-    #     #print(self.query.alias_refcount)
-    #     return super(TenantQuerySet,self)._update(values)
-    
-    #This API is called when there is a subquery. Injected tenant_ids for the subqueries.
-    def _as_sql(self, connection):
-        self.add_tenant_filters_with_joins()
-        return super(TenantQuerySet,self)._as_sql(connection)
-
-#Below is the manager related to the above class. 
-class TenantManager(TenantQuerySet.as_manager().__class__):
-    #Injecting tenant_id filters in the get_queryset.
-    #Injects tenant_id filter on the current model for all the non-join/join queries. 
-    def get_queryset(self):
-        current_tenant=get_current_tenant()
-        if current_tenant:
-            kwargs = { self.model.tenant_id: current_tenant.id}
-            return super(TenantManager, self).get_queryset().filter(**kwargs)
-        return super(TenantManager, self).get_queryset()
-
-#Abstract model which all the models related to tenant inherit.
-class TenantModel(models.Model):
-
-    #New manager from middleware
-    objects = TenantManager()
-    tenant_id=''
-
-    #adding tenant filters for save
-    #Citus requires tenant_id filters for update, hence doing this below change.
-    def _do_update(self, base_qs, using, pk_val, values, update_fields, forced_update):
-        current_tenant=get_current_tenant()
-        if current_tenant:
-            kwargs = { self.__class__.tenant_id: current_tenant.id}
-            base_qs = base_qs.filter(**kwargs)
-        return super(TenantModel,self)._do_update(base_qs, using, pk_val, values, update_fields, forced_update)
-
-    class Meta:
-        abstract = True
-
-
-class TenantForeignKey(models.ForeignKey):
-    '''
-    Should be used in place of models.ForeignKey for all foreign key relationships to
-    subclasses of TenantModel.
-
-    Adds additional clause to JOINs over this relation to include tenant_id in the JOIN
-    on the TenantModel.
-
-    Adds clause to forward accesses through this field to include tenant_id in the
-    TenantModel lookup.
-    '''
-
-    # Override
-    def get_extra_descriptor_filter(self, instance):
-        """
-        Return an extra filter condition for related object fetching when
-        user does 'instance.fieldname', that is the extra filter is used in
-        the descriptor of the field.
-
-        The filter should be either a dict usable in .filter(**kwargs) call or
-        a Q-object. The condition will be ANDed together with the relation's
-        joining columns.
-
-        A parallel method is get_extra_restriction() which is used in
-        JOIN and subquery conditions.
-        """
-        current_tenant = get_current_tenant()
-        if current_tenant:
-            return {instance.__class__.tenant_id: current_tenant.id}
-        else:
-            logger.warn('TenantForeignKey field %s.%s on instance "%s" '
-                        'accessed without a current tenant set. '
-                        'This may cause issues in a partitioned environment. '
-                        'Recommend calling set_current_tenant() before accessing '
-                        'this field.',
-                        self.model.__name__, self.name, instance)
-            return super(TenantForeignKey, self).get_extra_descriptor_filter(instance)
-
-    # Override
-    def get_extra_restriction(self, where_class, alias, related_alias):
-        """
-        Return a pair condition used for joining and subquery pushdown. The
-        condition is something that responds to as_sql(compiler, connection)
-        method.
-
-        Note that currently referring both the 'alias' and 'related_alias'
-        will not work in some conditions, like subquery pushdown.
-
-        A parallel method is get_extra_descriptor_filter() which is used in
-        instance.fieldname related object fetching.
-        """
-
-        # Fetch tenant column names for both sides of the relation
-        lhs_model = self.model
-        rhs_model = self.related_model
-        lhs_tenant_id = lhs_model.tenant_id
-        rhs_tenant_id = rhs_model.tenant_id
-
-        # Fetch tenant fields for both sides of the relation
-        lhs_tenant_field = lhs_model._meta.get_field(lhs_tenant_id)
-        rhs_tenant_field = rhs_model._meta.get_field(rhs_tenant_id)
-
-        # Get references to both tenant columns
-        lookup_lhs = lhs_tenant_field.get_col(related_alias)
-        lookup_rhs = rhs_tenant_field.get_col(alias)
-
-        # Create "AND lhs.tenant_id = rhs.tenant_id" as a new condition
-        lookup = lhs_tenant_field.get_lookup('exact')(lookup_lhs, lookup_rhs)
-        condition = where_class()
-        condition.add(lookup, 'AND')
-        return condition
-
-
-
-def get_current_user():
-    """
-    Despite arguments to the contrary, it is sometimes necessary to find out who is the current
-    logged in user, even if the request object is not in scope.  The best way to do this is 
-    by storing the user object in middleware while processing the request.
-    """
-    return getattr(_thread_locals, 'user', None)
-
-def get_model_by_db_table(db_table):
-    for model in apps.get_models():
-        if model._meta.db_table == db_table:
-            return model
-    else:
-        # here you can do fallback logic if no model with db_table found
-        raise ValueError('No model found with db_table {}!'.format(db_table))
-        # or return None
-
-def get_current_tenant():
-    """
-    To get the Tenant instance for the currently logged in tenant.
-    example:
-        tenant = get_current_tenant()
-    """
-    tenant = getattr(_thread_locals, 'tenant', None)
-
-    # tenant may not be set yet, if request user is anonymous, or has no profile,
-    # if not tenant:
-    #     set_tenant_to_default()
-    
-    return getattr(_thread_locals, 'tenant', None)
-
-
-# def set_tenant_to_default():
-#     """
-#     Sets the current tenant as per BASE_TENANT_ID.
-#     """
-#     # import is done from within the function, to avoid trouble 
-#     from models import Tenant, BASE_TENANT_ID
-#     set_current_tenant( Tenant.objects.get(id=BASE_TENANT_ID) )
-    
-
-def set_current_tenant(tenant):
-    setattr(_thread_locals, 'tenant', tenant)
-
-
-class ThreadLocals(object):
-    """Middleware that gets various objects from the
-    request object and saves them in thread local storage."""
-    def process_request(self, request):
-        _thread_locals.user = getattr(request, 'user', None)
-
-        # Attempt to set tenant
-        if _thread_locals.user and not _thread_locals.user.is_anonymous():
-            try:
-                profile = _thread_locals.user.get_profile()
-                if profile:
-                    _thread_locals.tenant = getattr(profile, 'tenant', None)
-            except:
-                raise ValueError(
-                    """A User was created with no profile.  For security reasons, 
-                    we cannot allow the request to be processed any further.
-                    Try deleting this User and creating it again to ensure a 
-                    UserProfile gets attached, or link a UserProfile 
-                    to this User.""")
